requirements.txt
asgiref==3.8.1
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1

